% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{connection}
\alias{connection}
\alias{connect}
\alias{disconnect}
\title{Database}
\usage{
connect()

disconnect(con)
}
\arguments{
\item{con}{Connection, as returned by \code{connect}.}
}
\description{
Manage database connections for covid.
}
\keyword{internal}
