% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(..., embed_url = "http://embed.org")
}
\arguments{
\item{...}{A series of options to be used inside the app.}

\item{embed_url}{Base URL to embed chart.}
}
\description{
Run the Shiny Application
}
